Events model and controllers unit tests completion .
